[nrf fromlist] net: zperf: Make shell dependency optional
Now that we a proper API, shell is just optional, so, make the
dependency optional by refactoring the code.

Also, add a build test combination in twister.

Upstream PR: https://github.com/zephyrproject-rtos/zephyr/pull/53336
