+++ subsys/net/lib/zperf/CMakeLists.txt
@@ -5,11 +5,14 @@ zephyr_library_named(zperf)
 zephyr_library_sources(
   zperf_common.c
   zperf_session.c
-  zperf_shell.c
   zperf_udp_receiver.c
   zperf_udp_uploader.c
   zperf_tcp_receiver.c
   zperf_tcp_uploader.c
+)
+
+zephyr_library_sources_ifdef(CONFIG_NET_SHELL
+  zperf_shell.c
 )
 
 zephyr_library_include_directories(
